First attempt to run chain tests in CI (ccl/linux)
