+++ ci/tests/requirements.txt
@@ -1,2 +1,3 @@
 jsonrpc_websocket
 requests
+async
