+++ ci/tests/start-node.bash
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-# set -x
+BASE="$(cd -P "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
 case $(uname -s) in
     Linux*)
@@ -18,8 +18,25 @@ case $(uname -s) in
         ;;
 esac
 
+case ${LISP} in
+    lispworks*)
+        ls=''
+        ;;
+    ccl*)
+        lisp_cli="$HOME/bin/ccl"
+        ;;
+    sbcl*)
+        lisp_cli="/usr/local/bin/sbcl --disable-debugger"
+        ;;
+    *)
+        echo "Unknown Lisp dialect: $LISP, falling back to CCL"
+        lisp_cli="$HOME/bin/ccl"
+        ;;
+esac
+
 start_timeout=30
 node_id=${1:-1}
+rc=0
 
 tempfoo=`basename $0`
 TMPFILE=`mktemp /tmp/${tempfoo}.XXXXXX` || exit 1
@@ -30,7 +47,7 @@ cat >$TMPFILE <<EOF
 EOF
 
 echo Starting node${node_id} ...
-tmux new-session -d -s node${node_id} "rlwrap ccl -l $TMPFILE"
+tmux new-session -d -s node${node_id} "$BASE/run-in-repl.bash $TMPFILE"
 
 ${timeout_cli} ${start_timeout} sh -c 'until nc -z $0 $1 2> /dev/null; do echo "Sleeping 1 sec..."; sleep 1; done' localhost $((3139+${node_id}))
 
@@ -38,5 +55,8 @@ if nc -z localhost $((3139+${node_id})) 2>/dev/null ; then
   echo "Ok!"
 else
   echo "Couldn't connect to node for ${start_timeout} seconds. Exiting..."
-  exit 1
+  rc=1
 fi
+
+rm $TMPFILE
+exit ${rc}
